test: wrapper 객체에 대한 테스트 추가 (기존의 테스트를 이동)
- 단위 테스트이기 때문에 검증 로직은 wrapper에서 진행하는 게 맞다고 판단하여 이동하였습니다.
