+++ src/test/java/racingcar/domain/CarTest.java
@@ -4,16 +4,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static racingcar.domain.constant.CarConstant.CAR_NAME_MAX_LENGTH;
-import static racingcar.enumType.ExceptionMessage.BLANK_MESSAGE;
-import static racingcar.enumType.ExceptionMessage.LENGTH_MESSAGE;
-import static racingcar.provider.TestProvider.*;
+import static racingcar.provider.TestProvider.createTestCar;
 
 public class CarTest {
 
@@ -22,32 +16,6 @@ public class CarTest {
     @BeforeEach
     void init() {
         testCar = createTestCar("pobi");
-    }
-
-    @ParameterizedTest
-    @DisplayName("Car 객체 생성 시 validation 정상적으로 작동한다.")
-    @ValueSource(strings = {"junpk", "jney", "pobi", "neo"})
-    void givenNormalCarName_thenSuccess(final String carName) {
-        assertThatCode(() -> Car.create(carName))
-                .doesNotThrowAnyException();
-    }
-
-    @ParameterizedTest
-    @DisplayName("5글자 초과의 이름이 들어왔을 경우 예외가 발생한다.")
-    @ValueSource(strings = {"junpak", "journey", "pobiconan", "neocat"})
-    void givenFiveOverLength_thenFail(final String carName) {
-        assertThatThrownBy(() -> Car.create(carName))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(String.format(LENGTH_MESSAGE.getValue(), CAR_NAME_MAX_LENGTH.getValue()));
-    }
-
-    @ParameterizedTest
-    @DisplayName("이름이 공백일 경우 예외가 발생한다.")
-    @EmptySource
-    void givenBlankName_thenFail(final String carName) {
-        assertThatThrownBy(() -> Car.create(carName))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(String.format(BLANK_MESSAGE.getValue(), "이름"));
     }
 
     @ParameterizedTest
